Servicio Reporte Completado. Servicio Solicitud en proceso...
- Cambios menores
  + Modelo Solicitud: Tipo de solicitud ahora es
    un booleano que inidca si es un Reclamo.

  + Modelo Reporte Ren: Ventas Totales ahora es
    double.

  + Modelo Reporte Inv: Lista de prodocutos con
    stock bajo ahora se inicializa al momento
    crear una instancia nueva.

- Nuevo
  + VentaDTO: Representa una Venta del
    microservicio Ventas.

  + ResponseDTO: Representa una respuesta del
    microservicio Ventas, formada por un mensaje y
    un objeto.

  + PedidoDTO: Representa un Pedido del
    micro servicio pedidos.

  + VentaClient: Para realizar peticiones al micro
    servicio Ventas.

  + UsuarioClint: Para realizar peticiones al
    microservicio Usuarios.

- Cambios
  + Controlador Solicitud: Se dividió en Soporte y
    Reclamo Controlador.

  + Client Pedido: Corriegido y actualizado.

+++ src/main/java/duoc/amaru/reportes/model/ReporteRendimiento.java
@@ -17,7 +17,7 @@ import lombok.NoArgsConstructor;
 @Table(name = "reporte_rendimiento")
 public class ReporteRendimiento extends Reporte {
     
-    private int ventasTotales;
+    private double ventasTotales;
     private double ticketPromedio;
     private double tasaConversion;
     private double margenBruto;
@@ -28,7 +28,7 @@ public class ReporteRendimiento extends Reporte {
                               Long idEmp,
                               String formato,
                               String periodo,
-                              int ventasTotales,
+                              double ventasTotales,
                               double ticketPromedio,
                               double tasaConversion,
                               double margenBruto,
